update for gar users and lti users
- added entity and repository to log gar users activity
- added repository for LtiUserConnection
