+++ src/Repository/ClassroomUserConnectionLogRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace User\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+class ClassroomUserConnectionLogRepository extends EntityRepository
+{
+   
+}
