+++ src/Repository/LtiUserConnectionRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace User\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+class LtiUserConnectionRepository extends EntityRepository
+{
+   
+}
